Basic Zombie terminado

+++ Assets/Scripts/MonoBehaviour/ScriptChar/Character.cs
@@ -9,7 +9,7 @@ public abstract class Character : MonoBehaviour
     public float initialHitPoints;                       //Valor inicial da vida
     
     public abstract void ResetCharacter();
-    public abstract IEnumerator DanoCaractere(int ammount, float interval);
+    public abstract IEnumerator DanoCaractere(float ammount, float interval);
 
     public virtual void KillCharacter()
     {
