+++ Assets/Scripts/MonoBehaviour/ScriptChar/Player.cs
@@ -29,7 +29,7 @@ public class Player : Character
 
     }
 
-    public override IEnumerator DanoCaractere(int ammount, float interval)
+    public override IEnumerator DanoCaractere(float ammount, float interval)
     {
     while (true)
      {
